feat(go.d/sd/nl): make timeout and interval configurable
